Random number generators / distributions with GSL
This is a first implementation of random number generators
and distributions from the GNU Scientific Library.

This follows a factory pattern to associate with integers representing
rng and rnd types to simple objects to get random numbers from.

TODO:
command line line options definition and parsing.
plug with main programs
add to makefile
debug ?
expand to support as most as possible the gsl

git-svn-id: svn+ssh://svn.ligforge.imag.fr/svnroot/ggen/trunk@18 0533444f-c493-4ece-a72f-de083fb4a2d6