Create the ggen_rng class, big changes to random options
Now it is possible to use rng without rnd : the graph generator does
need a distribution, only a random number generator.

This was done by creating a class ggen_rng, moving all gsl_rng stuff to
this class and reworking the way command line arguments were defined.

The programs are sync'd with the new interface.

Additional work is needed to have a code that is clean and easy to
understand : error handling and arguments checking. Some way for the
user to give names instead of number for the types is probably a good
idea.

git-svn-id: svn+ssh://svn.ligforge.imag.fr/svnroot/ggen/trunk@73 0533444f-c493-4ece-a72f-de083fb4a2d6