libification + new binary interface
Too much to say about this. Almost all the code is here
because I wanted to have only one binary and a library. This
is easier for unit testing and probably easier to add more functions

The fact is that I will comment all the changes for the release and
that's all.

All the work is not ready. In fact most of the ggen binary code should
be reworked to remove code duplicatas.
The extract property code is not ready at all.

git-svn-id: svn+ssh://svn.ligforge.imag.fr/svnroot/ggen/branches/unstable@140 0533444f-c493-4ece-a72f-de083fb4a2d6