Replace CMAKE by AUTOCONF
Cmake tries to solve problems we don't have and make several
things hard.
We found the autotools suite easier to integrate in our workflow.
This patch makes huge changes, mainly in the structure of the project.
Still several little details need to be worked out like proper
headers for the library but all in all it works quite well.

git-svn-id: svn+ssh://svn.ligforge.imag.fr/svnroot/ggen/branches/unstable@184 0533444f-c493-4ece-a72f-de083fb4a2d6