Convert to C project using Igraph (0.5.3)
Boost is an horrible mess. C code is cleaner and
less prone to errors.
Igraph is a LOT better in terms of interface (even if its not perfect)

Note: this is a BIG commit, but splitting it will be horrible.
For history the perarnau/wip/igraph branch is a more precise
in terms of changes history but the commit msgs are not of a great help.