[fix] use vertex name attribute when printing info
This commit makes public the code to compute the name
of a vertex (using special attributes or printf).
Code is separated in safe/unsafe version because of igraph special
handling of missing attributes.

Analysis code is changed to use new interface. This make lp and
others to give info on vertices using the good names.