add a vagrant file for a fast test environment
Vagrant is an open source tool to create and manage virtual machines.
It can be used to provide an automated way of setting up a virtual machine
containing a "standard" environment. This is very useful to provide people
with a base on which you can test commands, fill bug reports, and document
findings. It can also be a good base to build reproducible papers on.

We provide a Vagrantfile setting up a VM with ubuntu 14.04 LTS 64bit and the
necessary packages for ggen, as well as an up-to-date ggen install.