add sensible seed default for rng
While all users should provide a rng seed to gsl for practical applications,
this patch adds a sensible and somewhat portable default seed, so that two
successive runs of ggen do not generate the same graph. It might have been
confusing to users otherwise.

The seed is fetched from /dev/urandom, printed for reproducibility and we log
a warning.