add a graph option to *-property command line
While it might seem silly to use the command line and add or analyse a property
of size 1, it make the command line a bit more "complete", and provides users
with a consistent interface for their scripts.

The graph option is --graph, and is supported only for the `add-property` and
`analyse-property` commands.