[fix] multiple errors in longest-antichain
This fixes some issues with sorted float vectors in the antichain computation.
Note that this is part of a longstanding problem in how ggen interacts with
igraph. We have multiple instances of igraph_vector_t usage to store vertex and
edge indices. This is wrong, and creates issues regularly. The ggen code is due
for a good refactor on that front.